+++ src/app/shared/selector.util.ts
@@ -1,0 +1,29 @@
+import { createSelector, MemoizedSelector, Selector } from '@ngrx/store';
+import { hasValue } from './empty.util';
+
+/**
+ * Export a function to return a subset of the state by key
+ */
+export function keySelector<T, V>(parentSelector: Selector<any, any>, subState: string, key: string): MemoizedSelector<T, V> {
+  return createSelector(parentSelector, (state: T) => {
+    if (hasValue(state) && hasValue(state[subState])) {
+      return state[subState][key];
+    }
+    else {
+      return undefined;
+    }
+  });
+}
+/**
+ * Export a function to return a subset of the state
+ */
+export function subStateSelector<T, V>(parentSelector: Selector<any, any>, subState: string): MemoizedSelector<T, V> {
+  return createSelector(parentSelector, (state: T) => {
+    if (hasValue(state) && hasValue(state[subState])) {
+      return state[subState];
+    }
+    else {
+      return undefined;
+    }
+  });
+}
